moving logics to useEffect on Explore
